add test-utils crate, which has some macros to reduce redundant code in test-setup. Use the test-utils in ceremonies and communities pallet
