+++ Cargo.toml
@@ -5,5 +5,6 @@ members = [
     'communities',
     'ceremonies',
     'bazaar',
-    'sybil-gate'
+    'sybil-gate',
+    'test-utils'
 ]
