+++ ceremonies/Cargo.toml
@@ -107,8 +107,11 @@ rev = "a208da16"
 
 [dev-dependencies.sp-keyring]
 package = "sp-keyring"
-git = "https://github.com/paritytech/substrate.git" 
+git = "https://github.com/paritytech/substrate.git"
 rev = "a208da16"
+
+[dev-dependencies.test-utils]
+path = "../test-utils"
 
 [features]
 default = ["std"]
